[BE-17] feat: 스텝 모집 공고 올리는 기능 (#34)
* refactor: 엔티티 임베디드 타입으로 수정, 필요없는 필드 삭제

* feat: StaffRecruitment 만드는 mapper 추가

* feat: 근무일 기준 enum 추가, 엔티티 필드 수정

* refactor: request의 필드 타입 변경

* feat: StaffRecruitmentJob 만드는 mapper 추가

* feat: StaffRecruitmentQuestion 만드는 mapper 추가

* feat: StaffRecruitmentImage 만드는 mapper 추가.

* refactor: 엔티티 index 필드 타입 long -> int 로 수정

* refactor: 엔티티 필드 삭제 및 추가

* refactor: 엔티티 필드 변경으로 인한 수정사항들

* refactor: 메서드명 수정

* feat: 스텝 지원 공고 작성 api 구현

* refactor: 생성자 작성 대신 롬복 어노테이션 이용

* refactor: 엔티티 필드 수정으로 인한 dto 변환 작업 리펙터링

* refactor: 엔티티 필드 수정으로 인한 리펙터링

+++ src/main/java/guesthouse/staffrecruitment/domain/model/StaffRecruitmentImage.java
@@ -23,5 +23,5 @@ public class StaffRecruitmentImage {
     private String imageUrl;
 
     @Column(name = "image_index")
-    private Long index;
+    private int index;
 }
